% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micerData.R
\docType{data}
\name{biData}
\alias{biData}
\title{Example binary classification dataset.}
\format{
\describe{
\item{ref}{correct label}
\item{pred}{predicted label}
}
}
\description{
Example binary classification dataset. "Mine" is the positive case and
"Not Mine" is the background class. There are 178 samples from the "Mine" class and
4,822 samples from the "Not Mine" class. Class proportions are based on landscape proportions.
There are a total of 5,000 samples.
}
\references{
Maxwell, A.E., Bester, M.S., Guillen, L.A., Ramezan, C.A., Carpinello, D.J., Fan, Y.,
Hartley, F.M., Maynard, S.M. and Pyron, J.L., 2020. Semantic segmentation deep learning for extracting
surface mine extents from historic topographic maps. Remote Sensing, 12(24), p.4145.
}
\keyword{datasets}
