% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micerData.R
\docType{data}
\name{compareData}
\alias{compareData}
\title{Data for multiclass classification comparison}
\format{
\describe{
\item{ref}{correct label}
\item{rfRred}{Random forest prediction}
\item{dfRred}{Single decision tree prediction}
}
}
\description{
Example multiclass classification dataset with the following classes: "PFO", "PEM", "RLP",
and "Not". PFO = Palustrine Forested; PEM = Palustrine Emergent; RLP = River, Lake, Pond; Not = Not
Wetland. There are 600 examples from each class.
}
\references{
These data are unpublished
}
\keyword{datasets}
