% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micerData.R
\docType{data}
\name{mcData}
\alias{mcData}
\title{Example multiclass classification dataset}
\format{
\describe{
\item{ref}{correct label}
\item{pred}{predicted label}
}
}
\description{
Example multiclass classification dataset with the following classes: "Barren" (n=163),
"Forest" (n=20,807), "Impervious" (n=426), "Low Vegetation" (n=3,182), "Mixed Dev" (n=520),
and "Water" (n=200). There are a total of 25,298 samples.
}
\references{
Maxwell, A.E., Strager, M.P., Warner, T.A., Ramezan, C.A., Morgan, A.N. and Pauley, C.E.,
2019. Large-area, high spatial resolution land cover mapping using random forests, GEOBIA, and NAIP
orthophotography: Findings and recommendations. Remote Sensing, 11(12), p.1409.
}
\keyword{datasets}
