% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micer.R
\name{miceCompare}
\alias{miceCompare}
\title{Statistically compare two models using a paired t-test and bootstrap samples of the assessment results}
\usage{
miceCompare(ref, result1, result2, reps, frac)
}
\arguments{
\item{ref}{column of reference data as factor data type.}

\item{result1}{column of predictions as factor data type (first result to compare).}

\item{result2}{column of predictions as factor data type (second result to compare).}

\item{reps}{number of bootstrap replicates to use. Default is 200.}

\item{frac}{proportion of samples to include in each bootstrap sample. Default is 0.7.}
}
\value{
paired t-test results including t-statistic, degrees of freedom, p-value, 95% confidence interval, and mean difference
}
\description{
Statistically compare two models using a paired t-test and bootstrap samples of the assessment results
}
\examples{
miceCompare(ref=compareData$ref,
result1=compareData$rfPred,
result2=compareData$dtPred,
reps=1000,
frac=.7)
}
