% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micerData.R
\docType{data}
\name{compareData}
\alias{compareData}
\title{Data for multiclass classification comparison}
\format{
\describe{
\item{ref}{correct label}
\item{rfRred}{random forest prediction}
\item{dfRred}{single decision tree prediction}
}
}
\description{
Example multiclass classification dataset with the following wetland-related classes:
"PFO", "PEM", "RLP", and "Not". PFO = Palustrine Forested; PEM = Palustrine Emergent;
RLP = River, Lake, Pond; Not = Not Wetland. There are 600 examples from each class relative to the
reference labels.
}
\references{
These data are unpublished
}
\keyword{datasets}
